requirements.txt
pandas
numpy
tqdm
litellm
scikit-learn
faiss-cpu
pyyaml
sentence-transformers
rank-bm25